+++ cutlistoptimizerservice/src/main/java/com/pedroedrasousa/cutlistoptimizer/comparator/SolutionBiggestUnusedTileAreaComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class SolutionBiggestUnusedTileAreaComparator implements Comparator<Solution> {
     @Override
     public int compare(Solution o1, Solution o2) {
-        return o2.getBiggestArea() - o1.getBiggestArea();
+        return (int)(o2.getBiggestArea() - o1.getBiggestArea());
     }
 }
